refac: move agent.py to agent/static.py
