+++ core/agent/static.py
