+++ test/mocks/agent.py
@@ -1,8 +1,2 @@
-from typing import Tuple
 
-import numpy as np
-
-from core.agent import Agent
-from core.base_types import ActType, ObsType
-from core.data_init import DataInitializer
